Adding pageable function & validation

+++ backend/src/main/java/com/example/cursos_backend/services/LessonService.java
@@ -3,14 +3,17 @@ package com.example.cursos_backend.services;
 import com.example.cursos_backend.dtos.LessonRequestDTO;
 import com.example.cursos_backend.dtos.LessonResponseDTO;
 import com.example.cursos_backend.exceptions.ValueNotFoundException;
+import com.example.cursos_backend.infra.AuthorizationHelper;
 import com.example.cursos_backend.model.Course;
 import com.example.cursos_backend.model.Lesson;
+import com.example.cursos_backend.model.User;
 import com.example.cursos_backend.repositories.CourseRepository;
 import com.example.cursos_backend.repositories.LessonRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -18,11 +21,14 @@ public class LessonService {
 
     private final LessonRepository lessonRepository;
     private final CourseRepository courseRepository;
+    private final AuthorizationHelper authorizationHelper;
 
-    public LessonResponseDTO createLesson(LessonRequestDTO request, Long courseId) {
+    public LessonResponseDTO createLesson(LessonRequestDTO request, Long courseId, User user) {
 
         Course course = courseRepository.findById(courseId)
                 .orElseThrow(() -> new ValueNotFoundException("Curso não encontrado"));
+
+        authorizationHelper.checkOwnerOrAdmin(course.getInstructor().getId(), user);
 
         Lesson lesson = new Lesson();
         lesson.setTitle(request.title());
@@ -36,19 +42,19 @@ public class LessonService {
         return toResponseDTO(saved);
     }
 
-    public List<LessonResponseDTO> getLessonByCourse(Long courseId) {
+    public Page<LessonResponseDTO> getLessonByCourse(Long courseId, Pageable pageable) {
 
-        List<Lesson> lessons = lessonRepository.findByCourseId(courseId);
+        Page<Lesson> lessons = lessonRepository.findByCourseId(courseId, pageable);
 
-        return lessons.stream()
-                .map(this::toResponseDTO)
-                .toList();
+        return lessons.map(this::toResponseDTO);
     }
 
-    public void deleteLesson (Long lessonId) {
+    public void deleteLesson(Long lessonId, User user) {
 
         Lesson lesson = lessonRepository.findById(lessonId)
                 .orElseThrow(() -> new ValueNotFoundException("Aula não encontrada"));
+
+        authorizationHelper.checkOwnerOrAdmin(lesson.getCourse().getInstructor().getId(), user);
 
         lessonRepository.delete(lesson);
     }
